refactor searchbar into own componment
